Add files via upload
Initial Portfolio Website:
External CSS is a form of CSS which is used to add styling to multiple HTML pages at a time. It helps to design the layout of many HTML web pages simultaneously. The external CSS is always saved with the .css extension, and through this file, we can change the complete style of our HTML web page.